+++ app/build.gradle.kts
@@ -4,15 +4,15 @@ plugins {
 }
 
 android {
-    namespace = "com.example.snsappproject"
+    namespace = "kr.camp.sns"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.example.snsappproject"
+        applicationId = "kr.camp.sns"
         minSdk = 30
         targetSdk = 34
         versionCode = 1
-        versionName = "1.0"
+        versionName = "1.0.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -26,17 +26,22 @@ android {
             )
         }
     }
+    
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
     }
+
     kotlinOptions {
         jvmTarget = "1.8"
+    }
+
+    viewBinding {
+        enable = true
     }
 }
 
 dependencies {
-
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
