+++ app/src/androidTest/java/kr/camp/sns/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.snsappproject
+package kr.camp.sns
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
