+++ app/src/main/java/kr/camp/sns/MainActivity.kt
@@ -1,10 +1,7 @@
-package com.example.snsappproject
+package kr.camp.sns
 
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
